Temporary fix for AWS encoding bug...
The AWS Ruby SDK S3 write method mutates the data passed to it so it
becomes ascii (even if it was utf-8). The only way to resolve this in
the short term is to force encoding back to utf-8. We'll need a better
long term solution once we come to look at different language sites